rename emailCampaign files to fit kebab case convention
