+++ demo/api/src/app.module.ts
@@ -37,8 +37,8 @@ import { Config } from "./config/config";
 import { ConfigModule } from "./config/config.module";
 import { DamFile } from "./dam/entities/dam-file.entity";
 import { DamFolder } from "./dam/entities/dam-folder.entity";
-import { EmailCampaignContentBlock } from "./emailCampaign/blocks/emailCampaign-content.block";
-import { EmailCampaignContentScope } from "./emailCampaign/emailCampaign-content-scope";
+import { EmailCampaignContentBlock } from "./email-campaign/blocks/email-campaign-content.block";
+import { EmailCampaignContentScope } from "./email-campaign/email-campaign-content-scope";
 import { MenusModule } from "./menus/menus.module";
 import { StatusModule } from "./status/status.module";
