+++ demo/api/src/email-campaign/blocks/email-campaign-link.block.ts
