+++ demo/api/src/email-campaign/email-campaign-content-scope.ts
